Plot Personal Best (PB) Trends

Visualizes the trend of personal best times for specific running distances.

plot_pbs(
  pbs_df,
  activity_type = "Run",
  distance_meters = NULL,
  add_trend_line = TRUE
)

Arguments

pbs_df

A data frame from `calculate_pbs` containing PB data. Must contain columns: 'activity_date', 'time_seconds', 'distance_label', 'is_pb'.

activity_type

Type(s) of activities to display (e.g., "Run"). Default "Run".

distance_meters

Optional. Numeric vector of distances (meters) to filter. If NULL, all distances in pbs_df are plotted.

add_trend_line

Logical. Whether to add a trend line to the plot. Default TRUE.

Value

A ggplot object showing PB trends, faceted by distance if multiple are plotted.

Details

Plots the trend of best efforts for specified distances, highlighting new PBs using pre-calculated data from `calculate_pbs`.

Visualizes data from `calculate_pbs`. Points show best efforts; solid points mark new PBs. Y-axis is MM:SS. Users must first call `calculate_pbs` to generate the required data.

Examples

if (FALSE) { # \dontrun{
# First calculate PBs using calculate_pbs function
# pbs_data <- calculate_pbs(
#   activities_data = your_activities,
#   activity_type = "Run",
#   distances_m = c(1000, 5000, 10000)
# )
# 
# Then plot the PBs
# plot_pbs(pbs_data)
# 
# Plot only specific distances
# plot_pbs(pbs_data, distance_meters = c(5000, 10000))
# 
# Plot without trend line
# plot_pbs(pbs_data, add_trend_line = FALSE)
} # }